increase resolution of terrain mesh

+++ src/math/noise.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-float perlin_noise_2d(uint32_t x, uint32_t y);
+float perlin_noise_2d(float x, float y);
 
 #endif /* __math_noise_h__ */
